+++ reveal-ext-api/src/main/java/io/revealbi/sdk/ext/api/ICredentialRepository.java
@@ -13,5 +13,5 @@ public interface ICredentialRepository extends IRVAuthenticationResolver {
 	
 	Map<String, Object> getDataSourceCredentials(String userId, String dataSourceId) throws IOException;
 	void setDataSourceCredentials(String userId, String dataSourceId, String credentialsId) throws IOException;
-	void dataSourceDeleted(String userId, String dataSourceId, String provider) throws IOException;
+	void dataSourceDeleted(String userId, String dataSourceId, String provider, String uniqueIdentifier) throws IOException;
 }
